:rocket:
Added Source test.
Refactored classes.
