+++ index.js
@@ -1,7 +1,12 @@
 const Option = require('./src/option');
 const Log = require('./src/log');
-
+const Source = require('./src/source');
+const Reference = require('./src/reference');
+const Doc = require('./src/doc');
 module.exports = {
   Option,
   Log,
+  Source,
+  Reference,
+  Doc,
 };
