+++ package.json
@@ -18,6 +18,8 @@
   "homepage": "https://github.com/mr-doc/mr-doc-utils#readme",
   "dependencies": {
     "chalk": "^1.1.1",
+    "fs-extra": "^0.27.0",
+    "is-glob": "^2.0.1",
     "lodash": "^4.6.1",
     "sparkles": "^1.0.0"
   },
